# Conflicts:
#	DA_QuanLyCuaHangBanDT/src/main/java/com/poly/entity/Account.java
#	DA_QuanLyCuaHangBanDT/src/main/java/com/poly/entity/DonHang.java
#	DA_QuanLyCuaHangBanDT/src/main/java/com/poly/repository/ChiTietSanPhamDAO.java
#	DA_QuanLyCuaHangBanDT/src/main/java/com/poly/repository/DonHangDAO.java

+++ DA_QuanLyCuaHangBanDT/src/main/java/com/poly/entity/ChiTietDonHang.java
@@ -31,7 +31,7 @@ public class ChiTietDonHang {
     private Integer soLuong;
 
     @Column(name="Gia")
-    private Float gia;
+    private Double gia;
     
     @ManyToOne
     @JoinColumn(name = "Madh")
